refactor: move service interfaces to daemon
